requirements.txt
livekit-agents==1.2.17
livekit-plugins-silero==1.2.8
livekit-plugins-deepgram==1.0.0
livekit-plugins-groq==1.2.17
#python-dotenv>=0.19.0
#langchain>=0.0.350
#langchain-core>=0.3.75
#langchain-groq>=0.3.8
#faiss-cpu==1.7.3

# Environment and utility
python-dotenv==1.0.0
protobuf==5.27.0
requests>=2.31.0